Remove the contract using when we remove the last reference.
