+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Assume/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
