+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Attributes/ContractClass/Diagnostics.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <diagnostics id="WTG2004" message="This project does not use Code Contracts." severity="Warning">
-	<suppressId>CS8019</suppressId>
 	<diagnostic>
 		<location>Test0.cs: (5,2-47)</location>
 	</diagnostic>
