+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Attributes/ContractClass/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 namespace Foo
 {
