+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Attributes/ContractClassFor/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 namespace Foo
 {
