+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Attributes/ContractVerification/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
