+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Attributes/Pure/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
