+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/BulkFix/Result.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics.CodeAnalysis;
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
