+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/EndContractBlock/Result.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
