+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/EnsuresOnThrow/Result.cs
@@ -1,11 +1,10 @@
 using System;
-using System.Diagnostics.Contracts;
 
 public static class Bob
 {
 	public static void Method(out object value)
 	{
 		value = new object();
-		throw new System.InvalidOperationException();
+		throw new InvalidOperationException();
 	}
 }
