+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Requires/Events/Result.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics.Contracts;
 
 public class Bob
 {
