+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Requires/UnknowType/Diagnostics.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <diagnostics id="WTG2004" message="This project does not use Code Contracts." severity="Warning">
 	<suppressId>CS0246</suppressId>
-	<suppressId>CS8019</suppressId>
 	<diagnostic>
 		<location>Test0.cs: (8,3-66)</location>
 	</diagnostic>
