+++ WTG.Analyzers.Test/TestData/CodeContractsAnalyzer/Requires/Unknown/Result.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.Contracts;
 
 public class Bob : IBob
 {
